+++ src/commonAlias.ts
@@ -1,10 +1,10 @@
 import * as vscode from 'vscode';
 import { AliasNameBody } from './AliasNameBody';
-import { isNullOrEmpty } from './constants';
+import { IsWindows, isNullOrEmpty } from './constants';
 import { MyConfig, getConfig } from './dynamicConfig';
 import { TerminalType } from "./enums";
-import { enableColorAndHideCommandLine, outputInfoByDebugModeByTime, outputWarnByTime } from "./outputUtils";
-import { isWindowsTerminalOnWindows } from "./terminalUtils";
+import { enableColorAndHideCommandLine, outputInfoByDebugModeByTime, outputInfoQuietByTime, outputWarnByTime } from "./outputUtils";
+import { isToolExistsInPath, isWindowsTerminalOnWindows } from "./terminalUtils";
 import { getPowerShellName, replaceSearchTextHolder, replaceTextByRegex } from "./utils";
 
 const ShouldUseFunctionRegex = /\$\d\b|[ "']+?\$[\*@]|\$\{@(:\d+)?\}|\n\s+/; // Check $1 or $* or $@ or "${@}" or "${@:2}"
@@ -20,7 +20,7 @@ const TrimForLoopWhite = /(%[a-zA-Z]\s+in\s+\(')\s+/g;
 const GetPowerShellCommandHeadBodyRegex = /^(.*?)\b((?:pwsh|PowerShell)\s+(?:-Command\s+)?)"\s*(.+?)\s*"\s*$/s;
 
 export function replacePowerShellVarsForLinuxAlias(body: string): string {
-  return body.replace(/(?<!\\)(\$[a-z]\w+)/g, '\\$1');
+  return body.replace(/(?<!\\)(\$[a-zA-Z]\w+)/g, '\\$1');
 }
 
 export function replacePowerShellQuoteForLinuxAlias(windowsCmdBody: string): string {
@@ -44,13 +44,13 @@ function removeHeadSpacesInEachLine(body: string, addHead: string): string {
   return body;
 }
 
-function getCodeToReplaceHeadSpacesToTab(varName: string = 'rawBody'): string {
+function getCodeToReplaceHeadSpacesToTab(varName: string = 'rawBody', newLine: string = 'newLine'): string {
   const replaceTabTo = getConfig().ReplaceTabTo;
   if (replaceTabTo === '\t') {
     return '';
   }
-  const spacePattern = `^ {${replaceTabTo.length}}`;
-  return String.raw`$chTab=[char]9; $${varName} = [regex]::Replace($${varName}, '${spacePattern}', $chTab.ToString(), [System.Text.RegularExpressions.RegexOptions]::Multiline);`;
+  const spacePattern = String.raw`^(\t*) {${replaceTabTo.length}}`;
+  return String.raw`$chTab=([char]9).ToString(); $${varName} = [string]::Join($${newLine}, ($${varName} | msr -t '${spacePattern}' -o '\1\t' -g -1 -aPAC));`;
 }
 
 function trimAliasBody(body: string): string {
@@ -99,9 +99,9 @@ const CommonAliasMap: Map<string, string> = new Map<string, string>()
   .set('gpc', String.raw`git branch | msr -t "^\s*\*\s*(\S+).*" -o "git pull origin \1 $*" -XM & del-this-tmp-list`)
   .set('gph', String.raw`git branch | msr -t "^\s*\*\s*(\S+).*" -o "git push origin \1 $*" -XM`)
   .set('gpc-sm', String.raw`git rev-parse --abbrev-ref HEAD | msr -t "(.+)" -o "git pull origin \1 --no-recurse-submodules" -XM
-          & del-this-tmp-list & msr -z "git submodule sync && git submodule update --init" -t "&&" -o "\n" -PAC | msr -XM -V ne0`)
+          & del-this-tmp-list & msr -z "git submodule sync && git submodule update --init --no-fetch" -t "&&" -o "\n" -PAC | msr -XM -V ne0`)
   .set('gpc-sm-reset', String.raw`git rev-parse --abbrev-ref HEAD | msr -t "(.+)" -o "git pull origin \1 --no-recurse-submodules" -XM
-          && msr -z "git submodule sync && git submodule update --init && git submodule update -f" -t "&&" -o "\n" -PAC | msr -XM -V ne0
+          && msr -z "git submodule sync && git submodule update --init --no-fetch && git submodule update -f" -t "&&" -o "\n" -PAC | msr -XM -V ne0
           & del-this-tmp-list
           & git status`)
   .set('gca', String.raw`git commit --amend --no-edit $*`)
@@ -117,11 +117,11 @@ const CommonAliasMap: Map<string, string> = new Map<string, string>()
   .set('git-shallow-clone', String.raw`echo git clone --single-branch --depth 1 $* && git clone --single-branch --depth 1 $*`)
   .set('git-clean', String.raw`msr -z "git clean -xffd && git submodule foreach --recursive git clean -xffd" -t "&&" -o "\n" -PAC | msr -XM`)
   .set('git-sm-prune', String.raw`msr -XM -z "git prune" && msr -XMz "git submodule foreach git prune"`)
-  .set('git-sm-init', String.raw`msr -XMz "git submodule sync" && echo git submodule update --init $* | msr -XM & del-this-tmp-list & git status`)
+  .set('git-sm-init', String.raw`msr -XMz "git submodule sync" && echo git submodule update --init --no-fetch $* | msr -XM & del-this-tmp-list & git status`)
   .set('git-sm-reset', String.raw`msr -XMz "git submodule sync" && msr -XMz "git submodule init" && echo git submodule update -f $*
           | msr -XM & del-this-tmp-list & git status`)
   .set('git-sm-restore', String.raw`echo git restore . --recurse-submodules $* | msr -XM & del-this-tmp-list & git status`)
-  .set('git-sm-reinit', String.raw`msr -XM -z "git submodule deinit -f ." && msr -XM -z "git submodule update --init" & git status`)
+  .set('git-sm-reinit', String.raw`msr -XM -z "git submodule deinit -f ." && msr -XM -z "git submodule update --init --no-fetch" & git status`)
   .set('git-sm-update-remote', String.raw`msr -XMz "git submodule sync" && echo git submodule update --remote $* | msr -XM & git status`)
   .set('git-cherry-pick-branch-new-old-commits', String.raw`git log $1 | msr -b "^commit $2" -q "^commit $3" -t "^commit (\w+)" -o "\1" -M -C
           | msr -s "^:(\d+):" -n --dsc -t "^:\d+:(?:\d+:)?\s+(\w+)" -o "git cherry-pick \1" -X -V ne0 $4 $5 $6 $7 $8 $9`)
@@ -145,10 +145,13 @@ const CommonAliasMap: Map<string, string> = new Map<string, string>()
             Write-Host 'Please copy pure alias body in the function.' -ForegroundColor Red;
             return;
           }
-          $sep = [System.Environment]::NewLine.SubString([System.Environment]::NewLine.Length -1);
-          $newBody = [string]::Join($sep, $rawBody).Trim();
-          ${getCodeToReplaceHeadSpacesToTab('newBody')}
+          $newLine = ([char]10).ToString();
+          $newBody = [string]::Join($newLine, $rawBody).Trim();
+          ${getCodeToReplaceHeadSpacesToTab('newBody', 'newLine')}
           $jsonBody = $newBody | ConvertTo-Json;
+          if ($PSVersionTable.PSVersion.Major -lt 7) {
+            $jsonBody = $jsonBody.Replace('\u0026', '&').Replace('\u003e', '>');
+          }
           Set-Clipboard $jsonBody;
           $jsonBody;
           $message = 'Copied one-line body(length = ' + $jsonBody.Length + ') above to clipboard, you can paste it to aliasBody in msr.xxx.commonAliasNameBodyList in vscode settings.json';
@@ -530,6 +533,14 @@ WindowsAliasMap.forEach((body, name, _) => {
   body = trimAliasBody(body).replace(TrimMultilineRegex, ' ');
   WindowsAliasMap.set(name, body);
 });
+
+if (IsWindows) {
+  const [hasPwsh, path] = isToolExistsInPath('pwsh.exe', TerminalType.CMD);
+  if (hasPwsh) {
+    WindowsAliasMap.delete('pwsh');
+    outputInfoQuietByTime(`Remove alias 'pwsh' on Windows since found pwsh.exe at ${path}`);
+  }
+}
 
 export function getCommonAliasMap(terminalType: TerminalType, writeToEachFile: boolean): Map<string, string> {
   let cmdAliasMap = new Map<string, string>();